rename car specific tier selector
